Preliminary fixes for upgrading the MSVC toolchain
Fixes / suppresses a couple warnings raised by the updated MSVC and
silences all C++17 deprecation warnings since we can only fix them after
we update to use C++17.

Bug: dawn:824
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/75072
Auto-Submit: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Loko Kung <lokokung@google.com>

+++ src/tests/unittests/wire/WireAdapterTests.cpp
@@ -193,7 +193,7 @@ namespace {
                         for (wgpu::FeatureName feature : fakeFeatures) {
                             *(features++) = static_cast<WGPUFeatureName>(feature);
                         }
-                        return fakeFeatures.size();
+                        return static_cast<uint32_t>(fakeFeatures.size());
                     })));
 
                 api.CallAdapterRequestDeviceCallback(apiAdapter, WGPURequestDeviceStatus_Success,
@@ -259,7 +259,7 @@ namespace {
                         for (wgpu::FeatureName feature : fakeFeatures) {
                             *(features++) = static_cast<WGPUFeatureName>(feature);
                         }
-                        return fakeFeatures.size();
+                        return static_cast<uint32_t>(fakeFeatures.size());
                     })));
 
                 // The device was actually created, but the wire didn't support its features.
